Removed the logic from the component showlist and gave it its own file in the hooks folder. Also removed type for showearchform. wrote the architecture md.
